Fix clones of clones to not be missing objects from the turnip-subordinate.
init_repository with clone_from now manually constructs
turnip-subordinate to contain all objects and refs from the original
repo and its turnip-subordinate.

When libgit2 clones a repository, it copies the objects directory over
verbatim, including objects/info/alternates. That doesn't work too well
when an alternate has a relative path that's not valid outside the repo.

We also don't want it to copy the alternate configuration at all -- when
creating a turnip-subordinate for a clone of a clone, the new
subordinate should be a single repo that includes all of the objects
from the original and its subordinate. There's no built-in way to link
the objects from multiple repos into one, but luckily it's easy to do
ourselves.