Limit half-closed connection handling to clients that are done writing.
Previously we somewhat dodgily implemented both types of half-closure in
both directions, but half-closure is only actually needed to close the
git backend process' stdin. So implement it only on
PackProxyServerProtocol and PackBackendProtocol.

HTTP and SSH still lack the ability to forward closed write connections,
but this simplifies that work. And SSH connections no longer hang if the
backend dies without telling the client it's done, as the dropped
connection will cause disconnections all the way up the stack.