Fix shallow clones over HTTP. (LP: #1547141)
git upload-pack --stateless-rpc doesn't know that no more negotation is
coming in the initial deepen request, so shallow HTTP clones hang unless
stdin is closed. Half-close the connection once the body is sent to fix
this.

Also requires the half-close and stderr fixes from the previous two
commits to work around weirdness. But it got there in the end.