Use new bytes-capable checkRefPermissions protocol
Git ref paths may not be valid UTF-8, so we need to treat them as bytes.
We can't deal perfectly with non-UTF-8 refs - they won't get scanned and
so won't show up in the webservice API or the web UI - but we can at
least allow them to round-trip through Launchpad at the git level.  This
migrates to a new version of the checkRefPermissions protocol which can
cope with ref paths being bytes rather than text.

LP: #1517559