Stop returning repo_path argument from init_repo
Since init_repo always operates on its first argument and either returns
it unmodified or raises an exception, it makes more sense to stop
returning repo_path and adjust callers instead.  In particular, this
allows us to fix an UnboundLocalError in the case where init_repo raises
an exception in a way that's more clearly correct than it would
otherwise be.