Handle some faults from checkRefPermissions
This virtinfo method can return "not found" or "unauthorized" faults in
some cases involving unlucky timing.  Handle these legitimate cases by
treating them as equivalent to the user having no permissions on any of
the requested refs.

(Launchpad currently avoids returning the "unauthorized" fault in order
to avoid ugly failure cases, but it would be more elegant for it to be
able to rely on turnip to handle that fault appropriately.)