Fix handling of Unicode refs under LC_ALL=C
get_refs passed text to Repository.lookup_reference, which caused it to
try to encode it using Python's idea of the filesystem encoding.  If run
in a non-UTF-8 locale, this encoding would be something other than
UTF-8, and thus might fail.

Make sure to pass bytes to pygit2 to avoid this.