charm: Fix handling of rabbitmq relations
Avoid trying to start turnip-celery when its code and storage
prerequisites are potentially unavailable.

Rearrange how we request access to rabbitmq.  Using the 'available'
state provided by the rabbitmq interface means we can guarantee that the
broker URL will always be available at that point, simplifying
configuration.  To ensure that things work if the rabbitmq relation is
removed and re-added, we must explicitly clear any previous access
request so that the rabbitmq-server charm notices that the relation data
has changed.