Run "git count-objects -v" in the correct directory
get_repack_data works fine when called from a git hook, because they're
already run from the correct directory; but it breaks when called from a
turnip-set-symbolic-ref handler, because those don't run with the
repository path as their working directory.  Explicitly set the working
directory in the latter case to fix this.