Disable receive.autogc
We have automatic repack jobs now, so it probably makes more sense to
rely on those even if they're somewhat delayed rather than doing the
garbage-collection in-band.  At the moment we often see significant
delays at the end of a push which look like the result of "git
receive-pack" running "git gc --auto".