Add series name to deployment artifact
We'll want to upgrade off bionic at some point, but that needs to be
handled carefully: new Juju units will need to be deployed using
deployment artifacts built for the appropriate Python version.  To make
it easier to see what's going on, add the Ubuntu series name as a suffix
to the build label.  Deployment artifacts are built with and for
particular Ubuntu series, so it makes sense to annotate them.