charm: Drop resources
These were only ever for local development convenience, and it's usually
much easier to just fetch the resource from Swift.  Now that we're
trying to upload the charm to Charmhub, having a resource attached to it
is an active impediment, because Charmhub doesn't have a good way to
deal with optional resources; so upgrade to a version of the turnip-base
layer that drops the resource declaration, and remove the associated
build machinery.