Add stop commits to detect-merges
The `detect-merges` API can be very slow for repositories with deep
history, because as currently defined it has to walk through the whole
ancestry of a given target commit looking for "mainline" (first parent
only) commits that introduce a merge of any of the given source commits.
In particular, this often times out for Linux kernel repositories.

However, we can do much better.  Merge proposals are normally only
created in Launchpad once the target branch has been scanned, and we can
reasonably assume that when a merge proposal is created it won't yet
have been merged.  This means that Launchpad could tell turnip's
`detect-merges` API to stop walking through history whenever it
encounters the commit that it had previously scanned on the target
branch, and then we'd only have to look through the changes introduced
by a push rather than the entire history of the branch.  This should be
much less likely to time out.

LP: #1955040